html/browsers/history/the-history-interface/combination_history_003.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>combination_history_003(After calling of pushState and replaceState methods, check length.)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    try {
      var first;
      var second;
      var third;
      first = window.history.length;
      window.history.pushState(1, document.title, '?x=1');
      second = window.history.length;
      window.history.replaceState(2, document.title, '?x=2');
      third = window.history.length;

      if ((second - first) == 1 && second == third) {
        test(function() {
          assert_true(true, "combination test is failed.");
        });
      } else {
        test(function() {
          assert_unreached("combination test is failed.");
        });
      }
    } catch (e) {
      test(function() {
        assert_unreached("call of pushState method is failed.");
      });
    }
    </script>
  </body>
</html>
